`ImpEx`: Respect readonly state of the files
