+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Ignore case for type in the reference type (e.g. `groups(customer.uid)`) [#545](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/545)
 - Added `Execute` and `Validate` actions to the toolbar [#553](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/553)
 - Added `Select Statement` action to the toolbar [#556](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/556)
+- Respect readonly state of the files [#557](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/557)
 
 ### `CockpitNG` enhancements
 - Added code completion for AdvancedSearch `operator` parameter [#537](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/537)
