+++ resources/META-INF/lang/optional-impex.xml
@@ -1,6 +1,6 @@
 <!--
   ~ This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
-  ~ Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+  ~ Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
   ~
   ~ This program is free software: you can redistribute it and/or modify
   ~ it under the terms of the GNU Lesser General Public License as
@@ -132,19 +132,13 @@
             <add-to-group group-id="ProjectViewPopupMenu" anchor="before" relative-to-action="CompareTwoFiles"/>
         </group>
 
-        <group id="ImpExTableFormat.Actions" class="com.intellij.idea.plugin.hybris.tools.remote.action.ImpexActionsGroup">
-            <action id="hybris.impex.selectTable" class="com.intellij.idea.plugin.hybris.impex.tableFormatting.actions.ImpexTableSelectAction">
-                <keyboard-shortcut first-keystroke="control alt s" keymap="$default"/>
-            </action>
-
-            <add-to-group group-id="CodeMenu" anchor="last"/>
-            <add-to-group group-id="EditorPopupMenu" anchor="last"/>
-        </group>
-
         <action id="hybris.impex.openQuery" class="com.intellij.idea.plugin.hybris.impex.actions.ImpExOpenQueryAction"/>
         <action id="hybris.impex.openSettings" class="com.intellij.idea.plugin.hybris.impex.actions.ImpExOpenSettingsAction"/>
         <action id="hybris.impex.validate" class="com.intellij.idea.plugin.hybris.impex.actions.ImpExValidateAction"/>
         <action id="hybris.impex.import" class="com.intellij.idea.plugin.hybris.impex.actions.ImpExExecuteAction"/>
+        <action id="hybris.impex.selectTable" class="com.intellij.idea.plugin.hybris.impex.tableFormatting.actions.ImpexTableSelectAction">
+            <keyboard-shortcut first-keystroke="control alt s" keymap="$default"/>
+        </action>
 
         <group id="hybris.impex.table.actions">
             <reference id="hybris.impex.selectTable"/>
