+++ src/com/intellij/idea/plugin/hybris/impex/actions/AbstractImpExTableAction.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,7 +21,9 @@ import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
+import com.intellij.openapi.command.WriteCommandAction
 import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.editor.EditorModificationUtil
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiUtilBase
@@ -45,5 +47,21 @@ abstract class AbstractImpExTableAction : AnAction() {
         e.presentation.isEnabled = actionAllowed
     }
 
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+        val editor = e.getData(CommonDataKeys.EDITOR) ?: return
+        PsiUtilBase.getElementAtCaret(editor)
+            ?.let { getSuitableElement(it) }
+            ?.let {
+                WriteCommandAction.runWriteCommandAction(project) {
+                    if (EditorModificationUtil.requestWriting(editor)) {
+                        actionPerformed(project, editor, it)
+                    }
+                }
+            }
+    }
+
+    abstract fun actionPerformed(project: Project, editor: Editor, element: PsiElement)
+    abstract fun getSuitableElement(element: PsiElement): PsiElement?
     abstract fun isActionAllowed(project: Project, editor: Editor, element: PsiElement): Boolean
 }
