+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExTableRemoveAction.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -20,7 +20,6 @@ package com.intellij.idea.plugin.hybris.impex.actions
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexHeaderLine
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexValueLine
-import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
@@ -36,9 +35,15 @@ class ImpExTableRemoveAction : AbstractImpExTableAction() {
         }
     }
 
+    override fun actionPerformed(project: Project, editor: Editor, element: PsiElement) {
+        TODO("Not yet implemented")
+    }
+
+    override fun getSuitableElement(element: PsiElement): PsiElement? {
+        TODO("Not yet implemented")
+    }
+
     override fun isActionAllowed(project: Project, editor: Editor, element: PsiElement) = PsiTreeUtil
         .getParentOfType(element, ImpexValueLine::class.java, ImpexHeaderLine::class.java) != null
 
-    override fun actionPerformed(e: AnActionEvent) {
-    }
 }
